Rename mod "matcher" to "glob"
